rename bpe.py to train_bpe.py
